Verify version of VS installation
