+++ eng/models/NuGetInfo.cs
@@ -5,17 +5,14 @@
 using System;
 using Newtonsoft.Json;
 
-namespace Microsoft.DotNet.Framework.UpdateDependencies.Models
+namespace Microsoft.DotNet.Framework.Models
 {
-    public class LcuInfo
+    public class NuGetInfo
     {
         [JsonProperty(Required = Required.Always)]
-        public string OsVersion { get; set; } = String.Empty;
+        public string[] OsVersions { get; set; } = Array.Empty<string>();
 
         [JsonProperty(Required = Required.Always)]
-        public string[] RuntimeVersions { get; set; } = Array.Empty<string>();
-
-        [JsonProperty(Required = Required.Always)]
-        public string DownloadUrl { get; set; } = String.Empty;
+        public string NuGetClientVersion { get; set; } = String.Empty;
     }
 }
