+++ eng/update-dependencies/Microsoft.DotNet.Framework.UpdateDependencies.csproj
@@ -14,4 +14,8 @@
     <PackageReference Include="System.CommandLine" Version="0.1.0-e170603-2" />
     <PackageReference Include="System.Diagnostics.TextWriterTraceListener" Version="4.3.0" />
   </ItemGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\models\Microsoft.DotNet.Framework.Models.csproj" />
+  </ItemGroup>
 </Project>
