+++ tests/tests.sln
@@ -1,9 +1,11 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 15
-VisualStudioVersion = 15.0.27019.1
+# Visual Studio Version 16
+VisualStudioVersion = 16.0.29505.145
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Microsoft.DotNet.Framework.Docker.Tests", "Microsoft.DotNet.Framework.Docker.Tests/Microsoft.DotNet.Framework.Docker.Tests.csproj", "{4C892951-C119-4879-9EE9-EC48E56C5204}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Microsoft.DotNet.Framework.Docker.Tests", "Microsoft.DotNet.Framework.Docker.Tests\Microsoft.DotNet.Framework.Docker.Tests.csproj", "{4C892951-C119-4879-9EE9-EC48E56C5204}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Microsoft.DotNet.Framework.Models", "..\eng\models\Microsoft.DotNet.Framework.Models.csproj", "{D2524819-960D-41B4-AD9B-AF3FBAAE1951}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -15,6 +17,10 @@ Global
 		{4C892951-C119-4879-9EE9-EC48E56C5204}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{4C892951-C119-4879-9EE9-EC48E56C5204}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{4C892951-C119-4879-9EE9-EC48E56C5204}.Release|Any CPU.Build.0 = Release|Any CPU
+		{D2524819-960D-41B4-AD9B-AF3FBAAE1951}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{D2524819-960D-41B4-AD9B-AF3FBAAE1951}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{D2524819-960D-41B4-AD9B-AF3FBAAE1951}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{D2524819-960D-41B4-AD9B-AF3FBAAE1951}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
